Refactor CLI logic into separate module
Move CLI-related code from index.ts to src/main/lib/cli.ts:
- Launch directory parsing
- CLI install/uninstall functions
- isCliInstalled check
